fix int setter on stream map config

+++ pkg/axvdo/streamWrap.go
@@ -258,7 +258,7 @@ func VideoStreamConfigToVdoMap(cfg VideoSteamConfiguration) *VdoMap {
 	}
 	setIntIfNotNil := func(key string, value *int) {
 		if value != nil {
-			m.SetUint32(key, uint32(*value))
+			m.SetInt16(key, int16(*value))
 		}
 	}
 	setBoolIfNotNil := func(key string, value *bool) {
@@ -268,7 +268,7 @@ func VideoStreamConfigToVdoMap(cfg VideoSteamConfiguration) *VdoMap {
 	}
 
 	if cfg.Format != nil {
-		m.SetUint32("format", uint32(*cfg.Format))
+		m.SetInt16("format", int16(*cfg.Format))
 	}
 	setIntIfNotNil("buffer.count", cfg.BufferCount)
 	if cfg.BufferStrategy != nil {
@@ -294,17 +294,17 @@ func VideoStreamConfigToVdoMap(cfg VideoSteamConfiguration) *VdoMap {
 	setUint32IfNotNil("qp.p", cfg.Qpp)
 	setUint32IfNotNil("bitrate", cfg.Bitrate)
 	if cfg.RateControlMode != nil {
-		m.SetUint32("rc.mode", uint32(*cfg.RateControlMode))
+		m.SetInt16("rc.mode", int16(*cfg.RateControlMode))
 	}
 	if cfg.RateControlPriority != nil {
-		m.SetUint32("rc.prio", uint32(*cfg.RateControlPriority))
+		m.SetInt16("rc.prio", int16(*cfg.RateControlPriority))
 	}
 	setUint32IfNotNil("gop_length", cfg.GOPLength)
 	if cfg.H264Profile != nil {
-		m.SetUint32("h264.profile", uint32(*cfg.H264Profile))
+		m.SetInt16("h264.profile", int16(*cfg.H264Profile))
 	}
 	if cfg.H265Profile != nil {
-		m.SetUint32("h265.profile", uint32(*cfg.H265Profile))
+		m.SetInt16("h265.profile", int16(*cfg.H265Profile))
 	}
 	setUint32IfNotNil("zip.strength", cfg.ZipStrength)
 	setUint32IfNotNil("zip.max_gop_length", cfg.ZipMaxGOPLength)
@@ -320,7 +320,7 @@ func VideoStreamConfigToVdoMap(cfg VideoSteamConfiguration) *VdoMap {
 	setUint32IfNotNil("zip.min_fps_num", cfg.ZipMinFPSNum)
 	setUint32IfNotNil("zip.min_fps_den", cfg.ZipMinFPSDen)
 	if cfg.ZipProfile != nil {
-		m.SetUint32("zip.profile", uint32(*cfg.ZipProfile))
+		m.SetInt16("zip.profile", int16(*cfg.ZipProfile))
 	}
 	setUint32IfNotNil("abr.target_bitrate", cfg.AbrTarget_bitrate)
 	setUint32IfNotNil("abr.retention_time", cfg.AbrRetention_time)
